Ecommerce/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-mongodb-engine==0.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.2.1
Pillow==9.3.0
PyJWT==2.6.0
pymongo==3.12.3
pytz==2022.6
sqlparse==0.2.4
